requirements.txt
djangorestframework_simplejwt==5.5.0
├── Django [required: >=4.2, installed: 5.2.1]
│   ├── asgiref [required: >=3.8.1, installed: 3.8.1]
│   └── sqlparse [required: >=0.3.1, installed: 0.5.3]
├── djangorestframework [required: >=3.14, installed: 3.16.0]
│   └── Django [required: >=4.2, installed: 5.2.1]
│       ├── asgiref [required: >=3.8.1, installed: 3.8.1]
│       └── sqlparse [required: >=0.3.1, installed: 0.5.3]
└── PyJWT [required: >=1.7.1,<2.10.0, installed: 2.9.0]
drf-yasg==1.21.10
├── djangorestframework [required: >=3.10.3, installed: 3.16.0]
│   └── Django [required: >=4.2, installed: 5.2.1]
│       ├── asgiref [required: >=3.8.1, installed: 3.8.1]
│       └── sqlparse [required: >=0.3.1, installed: 0.5.3]
├── Django [required: >=2.2.16, installed: 5.2.1]
│   ├── asgiref [required: >=3.8.1, installed: 3.8.1]
│   └── sqlparse [required: >=0.3.1, installed: 0.5.3]
├── PyYAML [required: >=5.1, installed: 6.0.2]
├── inflection [required: >=0.3.1, installed: 0.5.1]
├── packaging [required: >=21.0, installed: 25.0]
├── pytz [required: >=2021.1, installed: 2025.2]
└── uritemplate [required: >=3.0.0, installed: 4.2.0]
pillow==11.3.0
pipdeptree==2.28.0
├── packaging [required: >=24.1, installed: 25.0]
└── pip [required: >=24.2, installed: 25.2]
psycopg2-binary==2.9.10
python-dotenv==1.1.1
python-magic==0.4.27
requests==2.32.4
├── charset-normalizer [required: >=2,<4, installed: 3.4.2]
├── idna [required: >=2.5,<4, installed: 3.10]
├── urllib3 [required: >=1.21.1,<3, installed: 2.5.0]
└── certifi [required: >=2017.4.17, installed: 2025.7.9]
tzdata==2025.2
whitenoise==6.9.0
